feat: implement package management features including adding, unbinding, and updating package end time for phone numbers

+++ app/src/main/java/org/LunaTelecom/dto/phone/GetNewNumberRequset.java
@@ -6,8 +6,8 @@ import jakarta.validation.constraints.PositiveOrZero;
 public class GetNewNumberRequset {
     @PositiveOrZero
     @NotBlank
-    public Long pool;
+    public long pool;
     @PositiveOrZero
     @NotBlank
-    public Long size;
+    public int size;
 }
